LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-focus-visible-with-mouse-expected.txt


option

FAIL Select should not match :focus-visible when using mouse. assert_equals: Nothing should be :focus-visible after opening select with mouse. expected null but got Element node <option>option</option>
